refactor: 랜덤 숫자 생성을 랜덤 볼 생성으로 개선
