--- src/main/java/lotto/generator/NumberGenerator.java
@@ -1,9 +1,0 @@
-package lotto.generator;
-
-import lotto.domain.Lotto;
-import lotto.domain.Lottos;
-
-public interface NumberGenerator {
-    Lotto generateLotto();
-    Lottos generateLottos(long lottoCount);
-}
